+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Spaceventure 0.9.4 (Jan 08, 2019)
 * Added new level
+* Small style improvements for switches and lasers
+* Conenction between a switch and a laser now has an animation when active
 
 # Spaceventure 0.9.3 (Jan 07, 2019)
 * New tutorial level